Controles de acceso

+++ routers/categorias.py
@@ -13,6 +13,10 @@ router = APIRouter(
 
 @router.post("/", response_model=schemas.Categoria, status_code=201)
 def crear_nueva_categoria(categoria: schemas.CategoriaCreate, db: Session = Depends(get_db), current_user: models.Usuario = Depends(get_current_user)):
+     # --- Verificación de Rol ---
+    if current_user.rol != "super-admin":
+        raise HTTPException(status_code=403, detail="No tienes permiso para crear categorías")
+   
     return crud.create_categoria(db=db, categoria=categoria)
 
 @router.get("/", response_model=List[schemas.Categoria])
@@ -29,6 +33,10 @@ def leer_categoria_por_id(categoria_id: int, db: Session = Depends(get_db)):
 
 @router.put("/{categoria_id}", response_model=schemas.Categoria)
 def actualizar_categoria(categoria_id: int, categoria: schemas.CategoriaCreate, db: Session = Depends(get_db), current_user: models.Usuario = Depends(get_current_user)):
+    # --- Verificación de Rol ---
+    if current_user.rol != "super-admin":
+        raise HTTPException(status_code=403, detail="No tienes permiso para actualizar categorías")
+    
     db_categoria = crud.update_categoria(db, categoria_id=categoria_id, categoria=categoria)
     if db_categoria is None:
         raise HTTPException(status_code=404, detail="Categoría no encontrada")
@@ -36,6 +44,10 @@ def actualizar_categoria(categoria_id: int, categoria: schemas.CategoriaCreate, 
 
 @router.delete("/{categoria_id}", response_model=schemas.Categoria)
 def borrar_categoria(categoria_id: int, db: Session = Depends(get_db), current_user: models.Usuario = Depends(get_current_user)):
+     # --- Verificación de Rol ---
+    if current_user.rol != "super-admin":
+        raise HTTPException(status_code=403, detail="No tienes permiso para borrar categorías")
+   
     db_categoria = crud.delete_categoria(db, categoria_id=categoria_id)
     if db_categoria is None:
         raise HTTPException(status_code=404, detail="Categoría no encontrada")
