+++ routers/equipos.py
@@ -13,6 +13,10 @@ router = APIRouter(
 
 @router.post("/", response_model=schemas.Equipo, status_code=201)
 def crear_nuevo_equipo(equipo: schemas.EquipoCreate, db: Session = Depends(get_db), current_user: models.Usuario = Depends(get_current_user)):
+     # --- Verificación de Rol ---
+    if current_user.rol not in ["super-admin", "editor"]:
+        raise HTTPException(status_code=403, detail="Permiso insuficiente")
+    
     return crud.create_equipo(db=db, equipo=equipo)
 
 @router.get("/", response_model=List[schemas.Equipo])
@@ -29,6 +33,10 @@ def leer_equipo_por_id(equipo_id: int, db: Session = Depends(get_db)):
 
 @router.put("/{equipo_id}", response_model=schemas.Equipo)
 def actualizar_equipo(equipo_id: int, equipo: schemas.EquipoCreate, db: Session = Depends(get_db), current_user: models.Usuario = Depends(get_current_user)):
+      # --- Verificación de Rol ---
+    if current_user.rol not in ["super-admin", "editor"]:
+        raise HTTPException(status_code=403, detail="Permiso insuficiente")
+   
     db_equipo = crud.update_equipo(db, equipo_id=equipo_id, equipo=equipo)
     if db_equipo is None:
         raise HTTPException(status_code=404, detail="Equipo no encontrado")
@@ -36,6 +44,10 @@ def actualizar_equipo(equipo_id: int, equipo: schemas.EquipoCreate, db: Session 
 
 @router.delete("/{equipo_id}", response_model=schemas.Equipo)
 def borrar_equipo(equipo_id: int, db: Session = Depends(get_db), current_user: models.Usuario = Depends(get_current_user)):
+    # --- Verificación de Rol ---
+    if current_user.rol not in ["super-admin", "editor"]:
+        raise HTTPException(status_code=403, detail="Permiso insuficiente")
+   
     db_equipo = crud.delete_equipo(db, equipo_id=equipo_id)
     if db_equipo is None:
         raise HTTPException(status_code=404, detail="Equipo no encontrado")
