+++ routers/eventos.py
@@ -12,6 +12,10 @@ router = APIRouter(
 
 @router.put("/{evento_id}", response_model=schemas.EventoPartido)
 def actualizar_evento(evento_id: int, evento: schemas.EventoPartidoCreate, db: Session = Depends(get_db), current_user: models.Usuario = Depends(get_current_user)):
+    # --- Verificación de Rol ---
+    if current_user.rol not in ["super-admin", "editor"]:
+        raise HTTPException(status_code=403, detail="Permiso insuficiente")
+   
     db_evento = crud.update_evento_partido(db, evento_id=evento_id, evento=evento)
     if db_evento is None:
         raise HTTPException(status_code=404, detail="Evento no encontrado")
@@ -19,6 +23,10 @@ def actualizar_evento(evento_id: int, evento: schemas.EventoPartidoCreate, db: S
 
 @router.delete("/{evento_id}", response_model=schemas.EventoPartido)
 def borrar_evento(evento_id: int, db: Session = Depends(get_db), current_user: models.Usuario = Depends(get_current_user)):
+    # --- Verificación de Rol ---
+    if current_user.rol not in ["super-admin", "editor"]:
+        raise HTTPException(status_code=403, detail="Permiso insuficiente")
+   
     db_evento = crud.delete_evento_partido(db, evento_id=evento_id)
     if db_evento is None:
         raise HTTPException(status_code=404, detail="Evento no encontrado")
